Create ninghth coding task
Write a program that takes an integer N and an array of N positive integers as input. Your program should output the position (0-based) of the array in which the value in that position has the maximum digit sum possible. A digit sum of some value is the sum of its digits. If there's more than one position that satisfies this condition, print the greatest one.

Note: The integer N will be given in the first line, and the array of N integers will be given in the second line (Each integer separated by a single space).